feat: refactory project structure
