+++ Dockerfile
@@ -2,6 +2,10 @@ FROM python:3.8-slim AS base
 
 WORKDIR /app
 
+RUN apt-get update && apt-get install -y --no-install-recommends \
+        libgl1 \
+        libglib2.0-0
+
 COPY Pipfile .
 RUN pip install pipenv
