+++ Pipfile
@@ -9,6 +9,7 @@ streamlit-ace = "==0.1.1"
 captcha = "==0.4"
 cvlib = "==0.2.7"
 tensorflow = "==2.9.1"
+opencv-python = "==4.6.0.66"
 
 [dev-packages]
 black = "==22.3.0"
