+++ app/pages.py
@@ -1,5 +1,10 @@
-from app.view.home import create
+import streamlit as st
+
+from app.view.form import registration
+from app.view.check_profile import check_profile
 
 
 def startup():
-    create()
+    st.title("Bank XPTO")
+    registration()
+    check_profile()
